+++ CHANGELOG.md
@@ -8,6 +8,11 @@ and this project adheres to
 
 ## [Unreleased]
 
+### Added
+
+- `mpvd time [file]` accepts an optional file path to print its duration via
+  `ffprobe` instead of the current track time position.
+
 ## [2.6.0] - 2026-09-23
 
 ### Added